requirements.txt
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.2
torch
tyro==0.8.3
wfcrl @ git+https://github.com/ifpen/wfcrl-env.git
stable-baselines3
wandb
seaborn
python-dotenv